Product quantity remaining now changes accurately with each transaction. Boolean success true, failure false.

+++ SA51_JavaCA/src/main/java/JavaCA/service/TransactionDetailsServiceImpl.java
@@ -7,8 +7,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import JavaCA.model.Product;
 import JavaCA.model.Transaction;
 import JavaCA.model.TransactionDetail;
+import JavaCA.repo.ProductRepository;
 import JavaCA.repo.TransactionDetailRepository;
 import JavaCA.repo.TransactionRepository;
 
@@ -22,19 +24,48 @@ public class TransactionDetailsServiceImpl implements TransactionDetailsService 
 	@Autowired
 	private TransactionDetailRepository transDRepo;
 	
+	@Autowired
+	private ProductRepository prodRepo;
+	
 	@Override
 	public ArrayList<TransactionDetail> findTransactionDetailsByProductId(long productId) {
 		return (ArrayList<TransactionDetail>) transDRepo.findTransactionDetailsByProductId(productId);
 	}
 	
 	@Override
-	public void saveTransactionDetail(TransactionDetail transactionDetail) {
-		transDRepo.save(transactionDetail);		
+	public boolean saveTransactionDetail(TransactionDetail transactionDetail) {
+		
+		int currentCount = transactionDetail.getProduct().getQuantity();
+		int qtyChange = transactionDetail.getQuantityChange();
+		if(transactionDetail.getTransactionType().toString() != "ORDER") {qtyChange = qtyChange * - 1;}
+		
+		if ((currentCount + qtyChange) >= 0){
+			Product p = transactionDetail.getProduct();
+			currentCount = currentCount + qtyChange;
+			p.setQuantity(currentCount);
+			prodRepo.save(p);
+			transDRepo.save(transactionDetail);	
+			return true;
+		}
+		return false;
 	}
 	
 	@Override
-	public void deleteTransactionDetail(TransactionDetail transactionDetail) {
-		transDRepo.delete(transactionDetail);
+	public boolean deleteTransactionDetail(TransactionDetail transactionDetail) {
+		
+		int currentCount = transactionDetail.getProduct().getQuantity();
+		int qtyChange = transactionDetail.getQuantityChange();
+		if(transactionDetail.getTransactionType().toString() != "ORDER") {qtyChange = qtyChange * -1;}
+		
+		if ((currentCount - qtyChange) >= 0){
+			Product p = transactionDetail.getProduct();
+			currentCount = currentCount - qtyChange;
+			p.setQuantity(currentCount);
+			prodRepo.save(p);
+			transDRepo.delete(transactionDetail);
+			return true;
+		}
+		return false;
 	}
 
 	@Override
